refactor: display filter values in the project list section
